refactor: resolve module path
