+++ packages/common/exceptions/not-implemented.exception.ts
@@ -1,6 +1,6 @@
 import { HttpException } from './http.exception';
 import { HttpStatus } from '../enums/http-status.enum';
-import { createHttpExceptionBody } from './../utils/http-exception-body.util';
+import { createHttpExceptionBody } from '../utils/http-exception-body.util';
 
 export class NotImplementedException extends HttpException {
   constructor(message?: string | object | any, error = 'Not Implemented') {
