+++ packages/common/exceptions/unprocessable-entity.exception.ts
@@ -1,6 +1,6 @@
 import { HttpException } from './http.exception';
 import { HttpStatus } from '../enums/http-status.enum';
-import { createHttpExceptionBody } from './../utils/http-exception-body.util';
+import { createHttpExceptionBody } from '../utils/http-exception-body.util';
 
 export class UnprocessableEntityException extends HttpException {
   constructor(message?: string | object | any, error = 'Unprocessable Entity') {
