+++ packages/common/files/interceptors/files.interceptor.ts
@@ -4,10 +4,10 @@ import { Inject, Optional } from '../../decorators';
 import { mixin } from '../../decorators/core/component.decorator';
 import { ExecutionContext } from '../../interfaces';
 import { MulterOptions } from '../../interfaces/external/multer-options.interface';
-import { NestInterceptor } from './../../interfaces/features/nest-interceptor.interface';
-import { MULTER_MODULE_OPTIONS } from './../files.constants';
-import { MulterModuleOptions } from './../interfaces';
-import { transformException } from './../multer/multer.utils';
+import { NestInterceptor } from '../../interfaces/features/nest-interceptor.interface';
+import { MULTER_MODULE_OPTIONS } from '../files.constants';
+import { MulterModuleOptions } from '../interfaces';
+import { transformException } from '../multer/multer.utils';
 
 type MulterInstance = any;
