+++ packages/common/interfaces/exceptions/rpc-exception-filter.interface.ts
@@ -1,5 +1,5 @@
 import { Observable } from 'rxjs';
-import { ArgumentsHost } from './../features/arguments-host.interface';
+import { ArgumentsHost } from '../features/arguments-host.interface';
 
 export interface RpcExceptionFilter<T = any, R = any> {
   catch(exception: T, host: ArgumentsHost): Observable<R>;
