+++ packages/common/interfaces/exceptions/ws-exception-filter.interface.ts
@@ -1,4 +1,4 @@
-import { ArgumentsHost } from './../features/arguments-host.interface';
+import { ArgumentsHost } from '../features/arguments-host.interface';
 
 export interface WsExceptionFilter<T = any> {
   catch(exception: T, host: ArgumentsHost);
