+++ packages/common/interfaces/features/execution-context.interface.ts
@@ -1,4 +1,4 @@
-import { Type } from './../index';
+import { Type } from '../index';
 import { ArgumentsHost } from './arguments-host.interface';
 
 export interface ExecutionContext extends ArgumentsHost {
