+++ packages/common/interfaces/modules/provider.interface.ts
@@ -1,4 +1,4 @@
-import { Type } from './../type.interface';
+import { Type } from '../type.interface';
 
 export type Provider =
   | Type<any>
