+++ packages/common/pipes/validation.pipe.ts
@@ -4,7 +4,7 @@ import { ValidatorOptions } from '../interfaces/external/validator-options.inter
 import { PipeTransform } from '../interfaces/features/pipe-transform.interface';
 import { loadPackage } from '../utils/load-package.util';
 import { isNil } from '../utils/shared.utils';
-import { Injectable } from './../decorators/core/component.decorator';
+import { Injectable } from '../decorators/core/component.decorator';
 
 export interface ValidationPipeOptions extends ValidatorOptions {
   transform?: boolean;
