+++ packages/common/test/pipes/parse-int.pipe.spec.ts
@@ -1,7 +1,7 @@
 import * as sinon from 'sinon';
 import { expect } from 'chai';
 import { ArgumentMetadata } from '../../interfaces';
-import { ParseIntPipe } from './../../pipes/parse-int.pipe';
+import { ParseIntPipe } from '../../pipes/parse-int.pipe';
 
 describe('ParseIntPipe', () => {
   let target: ParseIntPipe;
