+++ packages/common/test/pipes/validation.pipe.spec.ts
@@ -2,7 +2,7 @@ import * as sinon from 'sinon';
 import { expect } from 'chai';
 import { ArgumentMetadata } from '../../interfaces';
 import { IsString } from 'class-validator';
-import { ValidationPipe } from './../../pipes/validation.pipe';
+import { ValidationPipe } from '../../pipes/validation.pipe';
 
 class TestModel {
   constructor() {}
