+++ packages/common/test/utils/load-package.util.spec.ts
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import { loadPackage } from './../../utils/load-package.util';
+import { loadPackage } from '../../utils/load-package.util';
 
 describe('loadPackage', () => {
   describe('when package is available', () => {
