+++ packages/core/test/pipes/params-token-factory.spec.ts
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import { ParamsTokenFactory } from './../../pipes/params-token-factory';
+import { ParamsTokenFactory } from '../../pipes/params-token-factory';
 import { RouteParamtypes } from '../../../common/enums/route-paramtypes.enum';
 
 describe('ParamsTokenFactory', () => {
