+++ packages/websockets/test/context/exception-filters.context.spec.ts
@@ -2,7 +2,7 @@ import * as sinon from 'sinon';
 import { expect } from 'chai';
 import { UseFilters } from '../../../common/decorators/core/exception-filters.decorator';
 import { Catch } from '../../../common/decorators/core/catch.decorator';
-import { ExceptionFiltersContext } from './../../context/exception-filters-context';
+import { ExceptionFiltersContext } from '../../context/exception-filters-context';
 import { NestContainer } from '../../../core/injector/container';
 
 describe('ExceptionFiltersContext', () => {
