+++ packages/websockets/test/context/ws-proxy.spec.ts
@@ -1,6 +1,6 @@
 import * as sinon from 'sinon';
 import { expect } from 'chai';
-import { WsProxy } from './../../context/ws-proxy';
+import { WsProxy } from '../../context/ws-proxy';
 import { WsExceptionsHandler } from '../../exceptions/ws-exceptions-handler';
 import { WsException } from '../../exceptions/ws-exception';
